+++ src/libnnp/Element.h
@@ -36,6 +36,9 @@ class SymmetryFunctionGroup;
 class Element
 {
 public:
+    /** Default constructor
+     */
+    Element() {}
     /** Constructor using index.
      */
     Element(std::size_t const index, ElementMap const& elementMap);
@@ -43,7 +46,7 @@ public:
      *
      * Necessary because of #symmetryFunctions vector of pointers.
      */
-    ~Element();
+    virtual ~Element();
     /** Set #atomicEnergyOffset.
      */
     void                     setAtomicEnergyOffset(double atomicEnergyOffset);
@@ -114,7 +117,7 @@ public:
      * @param[in] statisticsLine Vector of strings containing statistics for
      *                           all symmetry functions.
      * @param[in] minS Minimum for scaling range.
-     * @param[in] maxS Minimum for scaling range.
+     * @param[in] maxS Maximum for scaling range.
      */
     void                     setScaling(
                                 SymmetryFunction::ScalingType   scalingType,
@@ -195,7 +198,7 @@ public:
     /// Symmetry function statistics.
     SymmetryFunctionStatistics statistics;
 
-private:
+protected:
     /// Copy of element map.
     ElementMap                            elementMap;
     /// Global index of this element.