+++ src/libnnpif/CabanaMD/typesCabana.h
@@ -1,0 +1,34 @@
+// n2p2 - A neural network potential package
+// Copyright (C) 2018 Andreas Singraber (University of Vienna)
+// Copyright (C) 2020 Saaketh Desai and Sam Reeve
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+#ifndef TYPESCABANAMD_H
+#define TYPESCABANAMD_H
+
+constexpr double CFLENGTH = 1.889726;
+constexpr double CFENERGY = 0.036749;
+constexpr double CFFORCE = CFLENGTH / CFENERGY;
+
+// Extracted from SymmetryFunction class
+enum ScalingType {
+    ST_NONE,
+    ST_SCALE,
+    ST_CENTER,
+    ST_SCALECENTER,
+    ST_SCALESIGMA
+};
+
+#endif // TYPES_H